Changed from CSS Grids to Flexbox and added messy js code
I changed from using css grids to flexbox based on the look of the
gallery being created with css grids. If I had a set of images that
were not random then this would be much easier and I would be able to
use css grids. This is not the case though and I will be making the
image gallery similar to that of Pinterest with the staggered images
based on height. I also cut the array being called from the Giphy API
into three different variables and then ran a for loop to sort them
into three columns. This was just a quick fix to test the flexbox
columns.